+++ src/main/java/com/naengo/api_server/domain/chat/entity/ChatRoom.java
@@ -20,10 +20,10 @@ public class ChatRoom {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "room_id")
-    private Long roomId;
+    private Integer roomId;
 
     @Column(name = "user_id", nullable = false)
-    private Long userId;
+    private Integer userId;
 
     @Column(nullable = false, length = 100)
     private String title;
