+++ src/main/java/com/naengo/api_server/domain/chat/repository/ChatMessageRepository.java
@@ -7,12 +7,12 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
+public interface ChatMessageRepository extends JpaRepository<ChatMessage, Integer> {
 
     @Query("""
            SELECT m FROM ChatMessage m
            WHERE m.roomId = :roomId
            ORDER BY m.createdAt ASC, m.messageId ASC
            """)
-    List<ChatMessage> findByRoomIdOrderByCreatedAt(@Param("roomId") Long roomId);
+    List<ChatMessage> findByRoomIdOrderByCreatedAt(@Param("roomId") Integer roomId);
 }
