+++ src/main/java/com/naengo/api_server/domain/recipe/entity/RecipeStep.java
@@ -19,7 +19,7 @@ public class RecipeStep {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "step_id")
-    private Long stepId;
+    private Integer stepId;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "recipe_id", nullable = false)
